Rename useParams to useRouteParams
