+++ packages/hydrogen/src/foundation/Route/Route.server.tsx
@@ -2,7 +2,7 @@ import React, {cloneElement, ReactElement} from 'react';
 import {useServerRequest} from '../ServerRequestProvider';
 import {matchPath} from '../../utilities/matchPath';
 import {Boomerang} from '../Boomerang/Boomerang.client';
-import {RouteParamsProvider} from '../useParams/RouteParamsProvider.client';
+import {RouteParamsProvider} from '../useRouteParams/RouteParamsProvider.client';
 
 export type RouteProps = {
   /** The URL path where the route exists. The path can contain variables. For example, `/products/:handle`. */
