+++ packages/hydrogen/src/foundation/useRouteParams/useRouteParams.tsx
@@ -2,9 +2,9 @@ import {useEnvContext} from '../ssr-interop';
 import {RouteParamsContext} from './RouteParamsProvider.client';
 
 /**
- *  The `useParams` hook retrieves the parameters of an active route. The hook is available in both server and client components.
+ *  The `useRouteParams` hook retrieves the parameters of an active route. The hook is available in both server and client components.
  */
-export function useParams() {
+export function useRouteParams() {
   const router = useEnvContext((req) => req.ctx.router, RouteParamsContext);
   return router.routeParams;
 }
